+++ packages/component-library/src/core/DataTable/cells/index.ts
@@ -1,0 +1,7 @@
+export { TextNumberCell } from './TextNumberCell';
+export { SelectCell } from './SelectCell';
+export { BinaryCell } from './BinaryCell';
+export { SimpleTextInput } from './SimpleTextInput';
+export { SimpleSelectInput } from './SimpleSelectInput';
+export { SimpleBinaryInput } from './SimpleBinaryInput';
+export type { TextNumberCellProps, SelectCellProps, BinaryCellProps, BaseCellProps } from './types';