+++ packages/component-library/src/core/HorizontalStackedBar/index.ts
@@ -1,0 +1,3 @@
+export * from './HorizontalBar';
+export * from './Legend';
+export * from './HorizontalStackedBar';